--- modules.py
@@ -1,605 +1,0 @@
-import os
-print(os.getcwd())
-
-import math
-import logging
-from random import random
-from functools import partial
-from pathlib import Path
-
-import torch
-from torch import nn, Tensor, einsum, IntTensor, FloatTensor, BoolTensor
-from torch.nn import Module
-import torch.nn.functional as F
-from torch.cuda.amp import autocast
-from torch.nn.utils import weight_norm
-
-from beartype import beartype
-from beartype.typing import Tuple, Optional, List, Union
-
-from einops.layers.torch import Rearrange
-from einops import rearrange, repeat, reduce, pack, unpack
-
-# from voicebox_pytorch.attend import Attend
-from attend import Attend
-from gateloop_transformer import SimpleGateLoopLayer as GateLoop
-
-import numpy
-import matplotlib.pyplot as plt
-
-LOGGER = logging.getLogger(__file__)
-logging.basicConfig(filename='model_debug.log', level=logging.INFO)
-
-# helper functions
-
-def exists(val):
-    return val is not None
-
-def identity(t):
-    return t
-
-def default(val, d):
-    return val if exists(val) else d
-
-def divisible_by(num, den):
-    return (num % den) == 0
-
-def is_odd(n):
-    return not divisible_by(n, 2)
-
-def coin_flip():
-    return random() < 0.5
-
-def pack_one(t, pattern):
-    return pack([t], pattern)
-
-def unpack_one(t, ps, pattern):
-    return unpack(t, ps, pattern)[0]
-
-def dynamic_range_decompression_torch(x, C=1):
-    return torch.exp(x) / C
-
-def dynamic_range_compression_torch(x, C=1, clip_val=1e-5):
-    return torch.log(torch.clamp(x, min=clip_val) * C)
-
-def spectral_normalize_torch(magnitudes):
-    output = dynamic_range_compression_torch(magnitudes)
-    return output
-
-
-
-# sinusoidal positions
-
-class LearnedSinusoidalPosEmb(Module):
-    """ used by @crowsonkb """
-
-    def __init__(self, dim):
-        super().__init__()
-        assert divisible_by(dim, 2)
-        half_dim = dim // 2
-        self.weights = nn.Parameter(torch.randn(half_dim))
-
-    def forward(self, x):
-        x = rearrange(x, 'b -> b 1')
-        freqs = x * rearrange(self.weights, 'd -> 1 d') * 2 * math.pi
-        fouriered = torch.cat((freqs.sin(), freqs.cos()), dim = -1)
-        return fouriered
-
-# rotary positional embeddings
-# https://arxiv.org/abs/2104.09864
-
-class RotaryEmbedding(Module):
-    def __init__(self, dim, theta = 50000):
-        super().__init__()
-        inv_freq = 1.0 / (theta ** (torch.arange(0, dim, 2).float() / dim))
-        self.register_buffer("inv_freq", inv_freq)
-
-    @property
-    def device(self):
-        return self.inv_freq.device
-
-    @autocast(enabled = False)
-    @beartype
-    def forward(self, t: Union[int, Tensor]):
-        if not torch.is_tensor(t):
-            t = torch.arange(t, device = self.device)
-
-        t = t.type_as(self.inv_freq)
-        freqs = torch.einsum('i , j -> i j', t, self.inv_freq)
-        freqs = torch.cat((freqs, freqs), dim = -1)
-        return freqs
-
-def rotate_half(x):
-    x1, x2 = x.chunk(2, dim = -1)
-    return torch.cat((-x2, x1), dim = -1)
-
-@autocast(enabled = False)
-def apply_rotary_pos_emb(pos, t):
-    return t * pos.cos() + rotate_half(t) * pos.sin()
-
-# convolutional positional generating module
-
-class ConvPositionEmbed(Module):
-    def __init__(
-        self,
-        dim,
-        *,
-        kernel_size,
-        groups = None
-    ):
-        super().__init__()
-        assert is_odd(kernel_size)
-        groups = default(groups, dim) # full depthwise conv by default
-
-        self.dw_conv1d = nn.Sequential(
-            nn.Conv1d(dim, dim, kernel_size, groups = groups, padding = kernel_size // 2),
-            nn.GELU()
-        )
-
-    def forward(self, x, mask = None):
-
-        if exists(mask):
-            mask = mask[..., None]
-            x = x.masked_fill(~mask, 0.)
-
-        x = rearrange(x, 'b n c -> b c n')
-        x = self.dw_conv1d(x)
-        out = rearrange(x, 'b c n -> b n c')
-
-        if exists(mask):
-            out = out.masked_fill(~mask, 0.)
-
-        return out
-
-# norms
-class RMSNorm(Module):
-    def __init__(
-        self,
-        dim
-    ):
-        super().__init__()
-        self.scale = dim ** 0.5
-        self.gamma = nn.Parameter(torch.ones(dim))
-
-    def forward(self, x):
-        return F.normalize(x, dim = -1) * self.scale * self.gamma
-
-class AdaptiveRMSNorm(Module):
-    def __init__(
-        self,
-        dim,
-        cond_dim = None
-    ):
-        super().__init__()
-        cond_dim = default(cond_dim, dim)
-        self.scale = dim ** 0.5
-
-        self.to_gamma = nn.Linear(cond_dim, dim)
-        self.to_beta = nn.Linear(cond_dim, dim)
-
-        # init to identity
-
-        nn.init.zeros_(self.to_gamma.weight)
-        nn.init.ones_(self.to_gamma.bias)
-
-        nn.init.zeros_(self.to_beta.weight)
-        nn.init.zeros_(self.to_beta.bias)
-
-    def forward(self, x, *, cond):
-        normed = F.normalize(x, dim = -1) * self.scale
-
-        gamma, beta = self.to_gamma(cond), self.to_beta(cond)
-        gamma, beta = map(lambda t: rearrange(t, 'b d -> b 1 d'), (gamma, beta))
-
-        return normed * gamma + beta
-
-# attention
-
-class MultiheadRMSNorm(Module):
-    def __init__(self, dim, heads):
-        super().__init__()
-        self.scale = dim ** 0.5
-        self.gamma = nn.Parameter(torch.ones(heads, 1, dim))
-
-    def forward(self, x):
-        return F.normalize(x, dim = -1) * self.gamma * self.scale
-
-class Attention(Module):
-    def __init__(
-        self, dim, dim_head = 64, heads = 8, dropout = 0, flash = False, qk_norm = False, qk_norm_scale = 10 ):
-        super().__init__()
-        self.heads = heads
-        dim_inner = dim_head * heads
-
-        scale = qk_norm_scale if qk_norm else None
-
-        self.attend = Attend(dropout, flash = flash, scale = scale)
-
-        self.qk_norm = qk_norm
-
-        if qk_norm:
-            self.q_norm = MultiheadRMSNorm(dim_head, heads = heads)
-            self.k_norm = MultiheadRMSNorm(dim_head, heads = heads)
-
-        self.to_qkv = nn.Linear(dim, dim_inner * 3, bias = False)
-        self.to_out = nn.Linear(dim_inner, dim, bias = False)
-
-    def forward(self, x, mask = None, rotary_emb = None):
-        h = self.heads
-
-        q, k, v = self.to_qkv(x).chunk(3, dim = -1)
-        q, k, v = map(lambda t: rearrange(t, 'b n (h d) -> b h n d', h = h), (q, k, v))
-
-        if self.qk_norm:
-            q = self.q_norm(q)
-            k = self.k_norm(k)
-
-        if exists(rotary_emb):
-            q, k = map(lambda t: apply_rotary_pos_emb(rotary_emb, t), (q, k))
-
-        out = self.attend(q, k, v, mask = mask)
-
-        out = rearrange(out, 'b h n d -> b n (h d)')
-        return self.to_out(out)
-
-# feedforward
-
-class GEGLU(Module):
-    def forward(self, x):
-        x, gate = x.chunk(2, dim = -1)
-        return F.gelu(gate) * x
-
-def FeedForward(dim, mult = 4, dropout = 0.):
-    dim_inner = int(dim * mult * 2 / 3)
-    return nn.Sequential(
-        nn.Linear(dim, dim_inner * 2),
-        GEGLU(),
-        nn.Dropout(dropout),
-        nn.Linear(dim_inner, dim)
-    )
-
-# transformer
-
-class Transformer(Module):
-    def __init__(
-        self,
-        dim,
-        *,
-        depth,
-        dim_head = 64,
-        heads = 8,
-        ff_mult = 4,
-        attn_dropout = 0.,
-        ff_dropout = 0.,
-        num_register_tokens = 0.,
-        attn_flash = False,
-        adaptive_rmsnorm = False,
-        adaptive_rmsnorm_cond_dim_in = None,
-        use_unet_skip_connection = False,
-        skip_connect_scale = None,
-        attn_qk_norm = False,
-        use_gateloop_layers = False,
-        gateloop_use_jax = False,
-    ):
-        super().__init__()
-        assert divisible_by(depth, 2)
-        self.layers = nn.ModuleList([])
-
-        self.rotary_emb = RotaryEmbedding(dim = dim_head)
-
-        self.num_register_tokens = num_register_tokens
-        self.has_register_tokens = num_register_tokens > 0
-
-        if self.has_register_tokens:
-            self.register_tokens = nn.Parameter(torch.randn(num_register_tokens, dim))
-
-        if adaptive_rmsnorm:
-            rmsnorm_klass = partial(AdaptiveRMSNorm, cond_dim = adaptive_rmsnorm_cond_dim_in)
-        else:
-            rmsnorm_klass = RMSNorm
-
-        self.skip_connect_scale = default(skip_connect_scale, 2 ** -0.5)
-
-        for ind in range(depth):
-            layer = ind + 1
-            has_skip = use_unet_skip_connection and layer > (depth // 2)
-
-            self.layers.append(nn.ModuleList([
-                nn.Linear(dim * 2, dim) if has_skip else None,
-                GateLoop(dim = dim, use_jax_associative_scan = gateloop_use_jax, post_ln = True) if use_gateloop_layers else None,
-                rmsnorm_klass(dim = dim),
-                Attention(dim = dim, dim_head = dim_head, heads = heads, dropout = attn_dropout, flash = attn_flash, qk_norm = attn_qk_norm),
-                rmsnorm_klass(dim = dim),
-                FeedForward(dim = dim, mult = ff_mult, dropout = ff_dropout)
-            ]))
-
-        self.final_norm = RMSNorm(dim)
-
-    @property
-    def device(self):
-        return next(self.parameters()).device
-
-    def forward(
-        self,
-        x,
-        mask = None,
-        adaptive_rmsnorm_cond = None
-    ):
-        batch, seq_len, *_ = x.shape
-
-        # add register tokens to the left
-
-        if self.has_register_tokens:
-            register_tokens = repeat(self.register_tokens, 'n d -> b n d', b = batch)
-
-            x, ps = pack([register_tokens, x], 'b * d')
-
-            if exists(mask):
-                mask = F.pad(mask, (self.num_register_tokens, 0), value = True)
-
-        # keep track of skip connections
-
-        skip_connects = []
-
-        # rotary embeddings
-
-        positions = seq_len
-
-        if self.has_register_tokens:
-            main_positions = torch.arange(seq_len, device = self.device, dtype = torch.long)
-            register_positions = torch.full((self.num_register_tokens,), -10000, device = self.device, dtype = torch.long)
-            positions = torch.cat((register_positions, main_positions))
-
-        rotary_emb = self.rotary_emb(positions)
-
-        # adaptive rmsnorm
-
-        rmsnorm_kwargs = dict()
-        if exists(adaptive_rmsnorm_cond):
-            rmsnorm_kwargs = dict(cond = adaptive_rmsnorm_cond)
-
-        # going through the attention layers
-
-        for skip_combiner, maybe_gateloop, attn_prenorm, attn, ff_prenorm, ff in self.layers:
-
-            # in the paper, they use a u-net like skip connection
-            # unclear how much this helps, as no ablations or further numbers given besides a brief one-two sentence mention
-
-            if not exists(skip_combiner):
-                skip_connects.append(x)
-            else:
-                skip_connect = skip_connects.pop() * self.skip_connect_scale
-                x = torch.cat((x, skip_connect), dim = -1)
-                x = skip_combiner(x)
-
-            if exists(maybe_gateloop):
-                x = maybe_gateloop(x) + x
-
-            attn_input = attn_prenorm(x, **rmsnorm_kwargs)
-            x = attn(attn_input, mask = mask, rotary_emb = rotary_emb) + x
-
-            ff_input = ff_prenorm(x, **rmsnorm_kwargs) 
-            x = ff(ff_input) + x
-
-        # remove the register tokens
-
-        if self.has_register_tokens:
-            _, x = unpack(x, ps, 'b * d')
-
-        return self.final_norm(x)
-
-##########################################################
-class ConvNeXtBlock(Module):
-    """ConvNeXt Block adapted from https://github.com/facebookresearch/ConvNeXt to 1D audio signal.
-
-    Args:
-        dim (int): Number of input channels.
-        intermediate_dim (int): Dimensionality of the intermediate layer.
-        layer_scale_init_value (float, optional): Initial value for the layer scale. None means no scaling.
-            Defaults to None.
-        adanorm_num_embeddings (int, optional): Number of embeddings for AdaLayerNorm.
-            None means non-conditional LayerNorm. Defaults to None.
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        intermediate_dim: int,
-        layer_scale_init_value: float,
-        hidden_dim: Optional[int] = None, # time embedding dim
-    ):
-        super().__init__()
-        self.dwconv = nn.Conv1d(dim, dim, kernel_size=7, padding=3, groups=dim)  # depthwise conv
-        self.adanorm = hidden_dim is not None
-        if hidden_dim:
-            self.norm = AdaLayerNorm(dim, hidden_dim ,eps=1e-6)
-        else:
-            self.norm = nn.LayerNorm(dim, eps=1e-6)
-        self.pwconv1 = nn.Linear(dim, intermediate_dim)  # pointwise/1x1 convs, implemented with linear layers
-        self.act = nn.GELU()
-        self.pwconv2 = nn.Linear(intermediate_dim, dim)
-        self.gamma = (
-            nn.Parameter(layer_scale_init_value * torch.ones(dim), requires_grad=True)
-            if layer_scale_init_value > 0
-            else None
-        )
-
-    def forward(self, x: torch.Tensor, cond: Optional[torch.Tensor] = None) -> torch.Tensor:
-        residual = x
-        x = self.dwconv(x)
-        x = x.transpose(1, 2)  # (B, C, T) -> (B, T, C)
-        if self.adanorm:
-            assert cond is not None
-            x = self.norm(x, cond)
-        else:
-            x = self.norm(x)
-        x = self.pwconv1(x)
-        x = self.act(x)
-        x = self.pwconv2(x)
-        if self.gamma is not None:
-            x = self.gamma * x
-        x = x.transpose(1, 2)  # (B, T, C) -> (B, C, T)
-
-        x = residual + x
-        return x
-
-class AdaLayerNorm(nn.Module):
-    """
-    Adaptive Layer Normalization module with learnable embeddings per `num_embeddings` classes
-
-    Args:
-        num_embeddings (int): Number of embeddings.
-        embedding_dim (int): Dimension of the embeddings.
-    """
-
-    # def __init__(self, num_embeddings: int, embedding_dim: int, eps: float = 1e-6):
-    def __init__(self, embedding_dim: int, hidden_dim: int, eps: float = 1e-6):
-        super().__init__()
-        self.eps = eps
-        self.dim = embedding_dim
-        # self.scale = nn.Embedding(num_embeddings=num_embeddings, embedding_dim=embedding_dim)
-        # self.shift = nn.Embedding(num_embeddings=num_embeddings, embedding_dim=embedding_dim)
-        self.scale = nn.Linear(hidden_dim, embedding_dim)
-        self.shift = nn.Linear(hidden_dim, embedding_dim)
-        nn.init.zeros_(self.scale.weight)
-        nn.init.ones_(self.scale.bias)
-        nn.init.zeros_(self.shift.weight)
-        nn.init.zeros_(self.shift.bias)
-        
-    # def forward(self, x: torch.Tensor, cond_embedding_id: torch.Tensor) -> torch.Tensor:
-    def forward(self, x: torch.Tensor, cond: torch.Tensor) -> torch.Tensor:
-        scale = self.scale(cond)
-        shift = self.shift(cond)
-        x = nn.functional.layer_norm(x, (self.dim,), eps=self.eps)
-        scale, shift = map(lambda t: t.unsqueeze(1).expand_as(x), (scale, shift))
-        x = x * scale + shift
-        return x
-
-
-class ResBlock1(nn.Module):
-    """
-    ResBlock adapted from HiFi-GAN V1 (https://github.com/jik876/hifi-gan) with dilated 1D convolutions,
-    but without upsampling layers.
-
-    Args:
-        dim (int): Number of input channels.
-        kernel_size (int, optional): Size of the convolutional kernel. Defaults to 3.
-        dilation (tuple[int], optional): Dilation factors for the dilated convolutions.
-            Defaults to (1, 3, 5).
-        lrelu_slope (float, optional): Negative slope of the LeakyReLU activation function.
-            Defaults to 0.1.
-        layer_scale_init_value (float, optional): Initial value for the layer scale. None means no scaling.
-            Defaults to None.
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        kernel_size: int = 3,
-        dilation: Tuple[int, int, int] = (1, 3, 5),
-        lrelu_slope: float = 0.1,
-        layer_scale_init_value: Optional[float] = None,
-    ):
-        super().__init__()
-        self.lrelu_slope = lrelu_slope
-        self.convs1 = nn.ModuleList(
-            [
-                weight_norm(
-                    nn.Conv1d(
-                        dim,
-                        dim,
-                        kernel_size,
-                        1,
-                        dilation=dilation[0],
-                        padding=self.get_padding(kernel_size, dilation[0]),
-                    )
-                ),
-                weight_norm(
-                    nn.Conv1d(
-                        dim,
-                        dim,
-                        kernel_size,
-                        1,
-                        dilation=dilation[1],
-                        padding=self.get_padding(kernel_size, dilation[1]),
-                    )
-                ),
-                weight_norm(
-                    nn.Conv1d(
-                        dim,
-                        dim,
-                        kernel_size,
-                        1,
-                        dilation=dilation[2],
-                        padding=self.get_padding(kernel_size, dilation[2]),
-                    )
-                ),
-            ]
-        )
-
-        self.convs2 = nn.ModuleList(
-            [
-                weight_norm(nn.Conv1d(dim, dim, kernel_size, 1, dilation=1, padding=self.get_padding(kernel_size, 1))),
-                weight_norm(nn.Conv1d(dim, dim, kernel_size, 1, dilation=1, padding=self.get_padding(kernel_size, 1))),
-                weight_norm(nn.Conv1d(dim, dim, kernel_size, 1, dilation=1, padding=self.get_padding(kernel_size, 1))),
-            ]
-        )
-
-        self.gamma = nn.ParameterList(
-            [
-                nn.Parameter(layer_scale_init_value * torch.ones(dim, 1), requires_grad=True)
-                if layer_scale_init_value is not None
-                else None,
-                nn.Parameter(layer_scale_init_value * torch.ones(dim, 1), requires_grad=True)
-                if layer_scale_init_value is not None
-                else None,
-                nn.Parameter(layer_scale_init_value * torch.ones(dim, 1), requires_grad=True)
-                if layer_scale_init_value is not None
-                else None,
-            ]
-        )
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        for c1, c2, gamma in zip(self.convs1, self.convs2, self.gamma):
-            xt = torch.nn.functional.leaky_relu(x, negative_slope=self.lrelu_slope)
-            xt = c1(xt)
-            xt = torch.nn.functional.leaky_relu(xt, negative_slope=self.lrelu_slope)
-            xt = c2(xt)
-            if gamma is not None:
-                xt = gamma * xt
-            x = xt + x
-        return x
-
-    def remove_weight_norm(self):
-        for l in self.convs1:
-            self.remove_weight_norm(l)
-            
-        for l in self.convs2:
-            self.remove_weight_norm(l)
-
-    @staticmethod
-    def get_padding(kernel_size: int, dilation: int = 1) -> int:
-        return int((kernel_size * dilation - dilation) / 2)
-
-
-def safe_log(x: torch.Tensor, clip_val: float = 1e-7) -> torch.Tensor:
-    """
-    Computes the element-wise logarithm of the input tensor with clipping to avoid near-zero values.
-
-    Args:
-        x (Tensor): Input tensor.
-        clip_val (float, optional): Minimum value to clip the input tensor. Defaults to 1e-7.
-
-    Returns:
-        Tensor: Element-wise logarithm of the input tensor with clipping applied.
-    """
-    return torch.log(torch.clip(x, min=clip_val))
-
-
-def symlog(x: torch.Tensor) -> torch.Tensor:
-    return torch.sign(x) * torch.log1p(x.abs())
-
-
-def symexp(x: torch.Tensor) -> torch.Tensor:
-    return torch.sign(x) * (torch.exp(x.abs()) - 1)
-
