+++ src/flowhigh/models/__init__.py
@@ -1,0 +1,2 @@
+from .melvoco import MelVoco
+from .flow import FLowHigh
