--- vocoder/BIGVGAN/bigvgan/alias_free_torch/__init__.py
@@ -1,6 +1,0 @@
-# Adapted from https://github.com/junjun3518/alias-free-torch under the Apache License 2.0
-#   LICENSE is in incl_licenses directory.
-
-from .filter import *
-from .resample import *
-from .act import *
