--- vocoder/BIGVGAN/config/bigvgan_48khz_256band_config.json
@@ -1,45 +1,0 @@
-{
-    "resblock": "1",
-    "num_gpus": 0,
-    "batch_size": 64,
-    "learning_rate": 0.0001,
-    "adam_b1": 0.8,
-    "adam_b2": 0.99,
-    "lr_decay": 0.999,
-    "seed": 1234,
-
-    "upsample_rates": [6,5,4,2,2],
-    "upsample_kernel_sizes": [12,11,8,4,4],
-    "upsample_initial_channel": 512,
-    "resblock_kernel_sizes": [3,7,11],
-    "resblock_dilation_sizes": [[1,3,5], [1,3,5], [1,3,5]],
-
-    "activation": "snakebeta",
-    "snake_logscale": true,
-
-    "resolutions": [[1024, 120, 600], [2048, 240, 1200], [512, 50, 240]],
-    "mpd_reshapes": [2, 3, 5, 7, 11],
-    "use_spectral_norm": false,
-    "discriminator_channel_mult": 1,
-
-    "segment_size": 15360,
-    "num_mels": 256,
-    "num_freq": 2049,
-    "n_fft": 2048,
-    "hop_size": 480,
-    "win_size": 2048,
-
-    "sampling_rate": 48000,
-
-    "fmin": 20,
-    "fmax": 24000,
-    "fmax_for_loss": null,
-
-    "num_workers": 4,
-
-    "dist_config": {
-        "dist_backend": "nccl",
-        "dist_url": "tcp://localhost:54321",
-        "world_size": 1
-    }
-}
